Now using floating point values for coordinate font coordinate calculation
Thought that it could have been a rounding issue that resulted in these weird offsets.
It wasn't. Still, this seems to be the better solution.

+++ core/src/de/bioviz/messages/MessageCenter.java
@@ -171,9 +171,9 @@ public class MessageCenter {
 				font.setColor(targetColor);
 				
 				TextBounds tb = font.getBounds(s.message);
-				int x = (int) ((s.x - tb.width / 2f) +
+				float x = ((s.x - tb.width / 2f) +
 							   Gdx.graphics.getWidth() / 2);
-				int y = (int) ((s.y + tb.height / 2f) +
+				float y = ((s.y + tb.height / 2f) +
 							   Gdx.graphics.getHeight() / 2);
 				if (s.size > 0) {
 					font.setScale(s.size / 100f);
